+++ packages/terra-framework-docs/CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Unreleased
 
 * Added
-  * Added an example and test file for a large submenu in the `terra-menu`.
+  * Added examples and tests for a large menu in the `terra-menu`.
 
 ## 1.67.0 - (February 22, 2024)